feat: allow vectoring per-slot overrides
Previously, the vector table was set up in a pseudo-direct mode where
all interrupts were redirected to a single `_start_trap` handler
regardless of number.

This change introduces 31 new `_start_trapN` targets that, by default,
retain the pseudo-direct behavior. However, they're weak symbols, so
they allow for individually overriding entries in the _vector_table, as
well as overriding the entire _vector_table itself.

This is in contrast with the earlier attempt in 9b77193 that jumped
straight to the corresponding `extern "C" interruptN`; instead, this
defers that work to the HAL and other downstream crates.